pyproject.toml
[project]
name = "arbo"
version = "0.2.0"
description = "Automated Polymarket Trading System — Reflexive Decay Harvester"
requires-python = ">=3.12"

dependencies = [
    # Async core
    "aiohttp>=3.9",

    # Database (PostgreSQL, NO Redis)
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",

    # Data validation & config
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",

    # Logging
    "structlog>=24.1",

    # Slack bot
    "slack-bolt>=1.27",

    # Polymarket
    "py-clob-client==0.34.6",

    # Blockchain (Layer 7: Order Flow)
    "web3>=7.0",

    # LLM (Gemini primary, Claude fallback) — used by Strategy B fallback
    "google-generativeai>=0.5",
    "anthropic>=0.40",

    # Web dashboard
    "fastapi>=0.109",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
    "python-multipart>=0.0.6",
    "psutil>=5.9",

    # Utilities
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.0",
    "pytest-cov>=4.1",
    "aioresponses>=0.7",
    "factory-boy>=3.3",
    "black>=24.1",
    "ruff>=0.2",
    "mypy>=1.8",
    "pre-commit>=3.6",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["arbo*"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"arbo/_archive/*.py" = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["arbo"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "aioresponses.*",
    "google.*",
    "py_clob_client.*",
    "web3.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["arbo/tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "integration: marks tests as integration tests (hit real APIs)",
]

[tool.coverage.run]
source = ["arbo"]
omit = ["arbo/tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
